"POMCP is not working with Settler of Catan":
	I try to work with POMCP to work with settler of Catan but it is not
working because of the at each iteration we can have a lot of action to
consider not take into account the observation which I took the state of
the game itself when I try to predict the card in other players hand. In
case of Java, my program yield the GC limited over which make me think
that the my hash table became too big and only some of will use at or
possibly really little will use.